refactor(AcademicCalendar): extract database connection out of AcademicCalendar
Justification:
AcademicCalendar should, logically, only be responsible
for the terms, courses and classes in the academic calendar itself. The
database connection is superflous and illogical. Further, AcademicCalendar
is tightly coupled with the University of Alberta LDAP server, which will
lead to difficult maintenance and extension in the future (i.e. connecting
to the course databases of other institutions).

New behaviour:
AcademicCalendar is initialized with a string representing
the institution name it will hold calendar data about. This string must
have a matching JSON config file in classtime/academic_databases/institutions
This config file must specify:
- the type of server connection
- all parameters necessary to instantiate a connection to that server type

See ualberta.json for an example config file

Institution server connections are to be created with the
AcademicDatabase.build factory method.

Closes #2