refactor(sched-scoring): build abstraction to allow for preferences, sliders, sort-by-options
Extract schedule scoring out to a dedicated class.
The scores are calculated on a per-test basis, and the overall
score is a sum of all those scores.

Weighting of each test is supported, and will likely be passed in
as an argument to calculate_scores() at a later date.
For now, weights are hardcoded defaults.