temp-fix(null-sections): keeps track of null sections found, and prints them when done
During the schedule generation, if a null section is found, it is added to the
class variable Schedule.NULL_SECTIONS, which is a set.

At the end of schedule generation, the set is printed.

Null sections are ignored, and are not used in ANY schedules.

This is a bandaid fix for invalid data issue (#30)