fix(scheduling): fill sections with timetable data, even when multiple timetable objects exist
Warnings will be logged when courses with:
- zero timetable objects, or
- multiple timetable objects
are fetched from the server.

You'll need to delete your database so that it can be refreshed
with the new data.

Either do `$ python manage.py delete_db`

Or do `$ rm /tmp/angular_flask.db`

This should (temporarily) fix #30. We'll need to keep watching
the log warnings to identify problem courses as they arise.