wip(section-dependencies): squashed bug. Sections fetch and save properly now.
Root cause was two-sided:
1. The 5-digit section identifier is named `class`, but that is a keyword in
python, so the primary key in the Section SQL table is `class_` instead. In
the generic `save()` method, the wrong primary key was used due to the
confusion. A better solution needs to be found for that.
2. The context-switching idiom `<obj>.use_<context/datatype>().do_<something>()`
which is used in AcademicCalendar and StandardLocalDatabase is a bug magnet.
Whenever using one of those two objects, it's possible for any method to have
the additional side effect of modifying the context.
This is bad, and bug-prone, and confusing, and issue #48 will improve it.