feat(sched-gen): diverse schedules, not all the same anymore
A diverse set of schedules is now returned.

After schedule generation, the list of schedules is "condensed" by
the following procedure:

- schedules are sorted by score, and by their timetable bitmap silhouette
- schedules are compared sequentially. If two schedules are similar
  enough, the second will be "folded" into the first by adding the second
  to the first's `more_like_this` list.

Schedules similarity is calculated as the fraction of the conflict
bitmaps which match identically.

The threshold for this fraction is defined as `Schedule.SIMILARITY_THRESHOLD`,
which is set to 0.950 at the time of this writing.

In the future, this can be improved by taking more factors into consideration
when calculating similarity.

Closes #66

"leave it better than you found it":

- more expressive and readable Schedule __repr__
- more expressive and readable schedule generation status report on return