electives: fix check/remove in accordion and drops in own group
Prior to this fix, drops in the same group posed a problem.  The first part of the change
was changing the drop zone from being each course-tile-row-thing to the entire group.  So now,
each group only has one drop zone on the container.  The second part of the change was to destroy
and recreate the dropped DOM node (and other nodes in its group) when dropping within its same group.
I do this by deep-copying the JS objects that become each DOM node, and then recreating.  This way, the
nodes are essentially reset to the normal position in their group!