fix: build issue where gulp seems to be overriding jQuery-UI
When jQuery-UI and jQuery are included in bower and build by gulp,
jQuery-UI is undefined when the angular-dragdrop lib tries to use it.
It seems that jQuery is being included twice? When it's included again
after UI, it overwrites the $ global and hence deletes UI.
I've tried to investigate why this is the case, but with no luck.

For now, remove from bower (include jQuery and jQuery-UI in the libs folder)