electives: transition from dialog view to full is consistant
This was an interesting one.

When the page is collapsed, the media query sets display: none for the
normal added tray.  I was using the same html template and same controller
for the dialog added tray - so the selectors used in animation were the same.
display: none doesn't actually remove from the DOM, it just hides and takes out
of the flow of the page.  So, when the dialog was being animated, the normal
added tray was also being affected.

The change is to make duplicates of the added controller and added html layout.
Duplicates, meaing the classnames are appended with "<name>2".  This way the
DOM elements are seperate, and stay consistant.